requirements.txt
aiofiles==24.1.0
cffi==1.17.1
contourpy==1.3.1
cryptography==41.0.3
cycler==0.12.1
fonttools==4.55.0
iniconfig==2.0.0
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
psutil==6.1.0
pycparser==2.22
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
setuptools==75.6.0
six==1.16.0
wheel==0.45.1
